requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
kombu==3.0.35
python-memcached==1.58
pytz==2016.6.1
six==1.10.0
